+++ src/main/java/com/ticketrush/api/controller/ReservationController.java
@@ -4,11 +4,11 @@ import com.ticketrush.application.reservation.model.ReservationCreateCommand;
 import com.ticketrush.application.reservation.model.ReservationLifecycleResult;
 import com.ticketrush.application.reservation.model.ReservationResult;
 import com.ticketrush.application.reservation.service.ReservationService;
-import com.ticketrush.domain.reservation.service.ReservationQueueService;
-import com.ticketrush.domain.reservation.service.AbuseAuditService;
-import com.ticketrush.domain.reservation.service.AdminRefundAuditService;
+import com.ticketrush.application.reservation.service.ReservationQueueService;
+import com.ticketrush.application.reservation.service.AbuseAuditService;
+import com.ticketrush.application.reservation.service.AdminRefundAuditService;
 import com.ticketrush.application.reservation.service.ReservationLifecycleService;
-import com.ticketrush.domain.reservation.service.SeatSoftLockService;
+import com.ticketrush.application.reservation.service.SeatSoftLockService;
 import com.ticketrush.domain.reservation.event.ReservationEvent;
 import com.ticketrush.domain.reservation.entity.AbuseAuditLog;
 import com.ticketrush.domain.reservation.entity.AdminRefundAuditLog;
