+++ src/main/java/com/ticketrush/application/reservation/service/AbuseAuditWriter.java
@@ -1,4 +1,4 @@
-package com.ticketrush.domain.reservation.service;
+package com.ticketrush.application.reservation.service;
 
 import com.ticketrush.domain.reservation.entity.AbuseAuditLog;
 import com.ticketrush.domain.reservation.repository.AbuseAuditLogRepository;
