+++ src/main/java/com/ticketrush/application/reservation/service/AdminRefundAuditService.java
@@ -1,4 +1,4 @@
-package com.ticketrush.domain.reservation.service;
+package com.ticketrush.application.reservation.service;
 
 import com.ticketrush.domain.reservation.entity.AdminRefundAuditLog;
 import com.ticketrush.domain.reservation.repository.AdminRefundAuditLogRepository;
