+++ src/main/java/com/ticketrush/application/reservation/service/ReservationLifecycleServiceImpl.java
@@ -11,8 +11,8 @@ import com.ticketrush.domain.reservation.port.outbound.ReservationSeatPort;
 import com.ticketrush.domain.reservation.port.outbound.ReservationUserPort;
 import com.ticketrush.domain.reservation.port.outbound.ReservationWaitingQueuePort;
 import com.ticketrush.domain.reservation.repository.ReservationRepository;
-import com.ticketrush.domain.reservation.service.AbuseAuditService;
-import com.ticketrush.domain.reservation.service.AdminRefundAuditService;
+import com.ticketrush.application.reservation.service.AbuseAuditService;
+import com.ticketrush.application.reservation.service.AdminRefundAuditService;
 import com.ticketrush.domain.user.User;
 import com.ticketrush.domain.user.UserRole;
 import com.ticketrush.global.config.ReservationProperties;
