+++ src/main/java/com/ticketrush/global/messaging/KafkaReservationConsumer.java
@@ -2,7 +2,7 @@ package com.ticketrush.global.messaging;
 
 import com.ticketrush.application.reservation.model.ReservationCreateCommand;
 import com.ticketrush.domain.reservation.event.ReservationEvent;
-import com.ticketrush.domain.reservation.service.ReservationQueueService;
+import com.ticketrush.application.reservation.service.ReservationQueueService;
 import com.ticketrush.application.reservation.service.ReservationService;
 import com.ticketrush.global.push.PushNotifier;
 import lombok.RequiredArgsConstructor;
