+++ src/test/java/com/ticketrush/application/reservation/service/ReservationLifecycleServiceIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.ticketrush.domain.reservation.service;
+package com.ticketrush.application.reservation.service;
 
 import com.ticketrush.application.reservation.service.ReservationLifecycleService;
 import com.ticketrush.application.reservation.service.ReservationLifecycleServiceImpl;
