+++ src/test/java/com/ticketrush/application/reservation/service/동시성_테스트_1_낙관적_락.java
@@ -1,5 +1,6 @@
-package com.ticketrush.domain.reservation.service;
+package com.ticketrush.application.reservation.service;
 
+import com.ticketrush.application.reservation.service.ReservationService;
 import com.ticketrush.domain.concert.entity.Concert;
 import com.ticketrush.domain.concert.entity.ConcertOption;
 import com.ticketrush.domain.concert.entity.Seat;
@@ -8,8 +9,7 @@ import com.ticketrush.domain.concert.service.ConcertService;
 import com.ticketrush.domain.reservation.repository.ReservationRepository;
 import com.ticketrush.domain.user.User;
 import com.ticketrush.domain.user.UserRepository;
-import com.ticketrush.global.lock.RedissonLockFacade;
-import com.ticketrush.api.dto.ReservationRequest;
+import com.ticketrush.application.reservation.model.ReservationCreateCommand;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -24,11 +24,25 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+/**
+ * <h2>동시성 테스트 - Step 1: 낙관적 락 (Optimistic Lock)</h2>
+ * <p>
+ * 본 테스트는 좌석 예약 시 발생할 수 있는 Race Condition 상황에서
+ * JPA의 {@code @Version}을 이용한 낙관적 락이 어떻게 데이터 정합성을 유지하는지 확인합니다.
+ * </p>
+ *
+ * <h3>검증 지표</h3>
+ * <ul>
+ *     <li><b>총 요청 수</b>: 30건</li>
+ *     <li><b>기대 성공 수</b>: 1건</li>
+ *     <li><b>기대 실패 수</b>: 29건 (Optimistic Lock 충돌)</li>
+ * </ul>
+ */
 @SpringBootTest
-class 동시성_테스트_3_분산_락 {
+class 동시성_테스트_1_낙관적_락 {
 
     @Autowired
-    private RedissonLockFacade redissonLockFacade;
+    private ReservationService reservationService;
 
     @Autowired
     private ReservationRepository reservationRepository;
@@ -48,13 +62,13 @@ class 동시성_테스트_3_분산_락 {
     @BeforeEach
     void setUp() {
         String suffix = String.valueOf(System.currentTimeMillis());
-        User user = userRepository.save(new User("testUser_Redis_" + suffix));
+        User user = userRepository.save(new User("testUser_" + suffix));
         targetUserId = user.getId();
 
         Concert concert = concertService.createConcert(
-                "testConcert_distributed_" + suffix,
-                "testArtist_distributed_" + suffix,
-                "testEntertainment_distributed_" + suffix
+                "testConcert_optimistic_" + suffix,
+                "testArtist_optimistic_" + suffix,
+                "testEntertainment_optimistic_" + suffix
         );
         ConcertOption option = concertService.addOption(concert.getId(), LocalDateTime.now().plusDays(1));
         concertService.createSeats(option.getId(), 1);
@@ -66,7 +80,7 @@ class 동시성_테스트_3_분산_락 {
     }
 
     @Test
-    @DisplayName("Redis 분산 락 동작 확인: 30명 동시 예약 시 1명만 성공 (Facade 적용 버전)")
+    @DisplayName("낙관적 락 동작 확인: 30명 동시 예약 시 1명만 성공")
     void concurrentTest() throws InterruptedException {
         int threadCount = 30;
         ExecutorService executorService = Executors.newFixedThreadPool(threadCount);
@@ -78,11 +92,10 @@ class 동시성_테스트_3_분산_락 {
         for (int i = 0; i < threadCount; i++) {
             executorService.submit(() -> {
                 try {
-                    // Facade를 통한 분산 락 예약 호출
-                    redissonLockFacade.createReservation(new ReservationRequest(targetUserId, targetSeatId));
+                    reservationService.createReservation(new ReservationCreateCommand(targetUserId, targetSeatId));
                     successCount.incrementAndGet();
                 } catch (Exception e) {
-                    System.out.println("예약 실패: " + e.getMessage());
+                    // ObjectOptimisticLockingFailureException 발생 예상
                     failCount.incrementAndGet();
                 } finally {
                     latch.countDown();
@@ -94,11 +107,11 @@ class 동시성_테스트_3_분산_락 {
 
         long actualReservationCount = reservationRepository.findBySeatId(targetSeatId).size();
         
-        System.out.println("================ 실험 결과 (Redis 분산 락 - Facade) ================");
-        System.out.println("성공 횟수: " + successCount.get());
-        System.out.println("실패 횟수: " + failCount.get());
+        System.out.println("================ 실험 결과 (낙관적 락) ================");
+        System.out.println("성공 유저 수: " + successCount.get());
+        System.out.println("실패 유저 수: " + failCount.get());
         System.out.println("해당 좌석 예약 건수: " + actualReservationCount);
-        System.out.println("===============================================================");
+        System.out.println("=========================================");
 
         assertThat(actualReservationCount).isEqualTo(1);
     }
