+++ src/test/java/com/ticketrush/application/reservation/service/동시성_테스트_4_비동기_대기열.java
@@ -1,4 +1,4 @@
-package com.ticketrush.domain.reservation.service;
+package com.ticketrush.application.reservation.service;
 
 import com.ticketrush.api.controller.ReservationController;
 import com.ticketrush.domain.concert.entity.Concert;
